// app/schemas/authSchema.ts
import z from "zod";

export const registerSchema = z.object({
  name: z.string().min(1, "Name is required"),
  email: z.string("Invalid email address"),
  password: z.string().min(8, "Password must be at least 8 characters long"),
});

export const signInSchema = z.object({
  email: z.string("Invalid email address"),
  password: z.string().min(8, "Password must be at least 8 characters long"),
});